Implementation of coloring the number of pumps around the ApplyOpenStyle method

+++ src/br/com/wallace/ms/model/Field.java
@@ -128,8 +128,8 @@ public class Field {
 		return protectedVariable || unraveledVariable;
 	}
 	
-	public long minesInTheNeighborhood() {
-		return neighbors.stream().filter(n -> n.undermine).count();
+	public int minesInTheNeighborhood() {
+		return (int) neighbors.stream().filter(n -> n.undermine).count();
 	}
 	
 	public void restart() {
